style: enforce consistent codestyle via ruff config and auto-fixes
- Configure ruff: isort, pyupgrade, bugbear, flake8 rules
- Auto-fix import sorting, redundant parens, exception chaining
- Remove from __future__ import annotations (Python 3.13+)
- Convert debug print() to logger calls
- Fix typo applicaiton → application in icon_resolver
- Remove bare returns in animator setters
- Move TYPE_CHECKING imports to runtime where used at runtime

+++ src/shared/widgets/customrevealer.py
@@ -113,7 +113,7 @@ class SlideRevealer(Gtk.Overlay):
         x, y = self._get_position_at_progress_cached(progress)
 
         # Round to nearest pixel for actual positioning
-        pixel_x, pixel_y = int(round(x)), int(round(y))
+        pixel_x, pixel_y = round(x), round(y)
         self._fixed.move(self.child, pixel_x, pixel_y)
         self.queue_draw()
